Version 2.0.0

+++ web/controllers/groups.go
@@ -1,18 +1,21 @@
-package pages
+package controllers
 
 import (
+	"errors"
 	"fmt"
 	"net/http"
 	"strconv"
 	"strings"
 
 	"github.com/ordinary-dev/phoenix/database"
+	"github.com/ordinary-dev/phoenix/web/sessions"
 )
 
 func CreateGroup(w http.ResponseWriter, r *http.Request) {
-	// Save new group to the database.
+	username := sessions.GetUsername(r.Context())
 	group := database.Group{
-		Name: strings.TrimSpace(r.FormValue("groupName")),
+		Name:     strings.TrimSpace(r.FormValue("groupName")),
+		Username: &username,
 	}
 
 	if err := database.CreateGroup(&group); err != nil {
@@ -20,7 +23,6 @@ func CreateGroup(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// This page is called from the settings, return the user back.
 	http.Redirect(w, r, fmt.Sprintf("/settings#group-%v", group.ID), http.StatusFound)
 }
 
@@ -31,13 +33,24 @@ func UpdateGroup(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	group, err := database.GetGroup(int(id))
+	if err != nil {
+		ShowError(w, http.StatusInternalServerError, err)
+		return
+	}
+
+	username := sessions.GetUsername(r.Context())
+	if group.Username == nil || *group.Username != username {
+		ShowError(w, http.StatusForbidden, errors.New("you are not the owner of the group"))
+		return
+	}
+
 	newName := strings.TrimSpace(r.FormValue("groupName"))
 	if err := database.UpdateGroup(int(id), newName); err != nil {
 		ShowError(w, http.StatusInternalServerError, err)
 		return
 	}
 
-	// This page is called from the settings, return the user back.
 	http.Redirect(w, r, fmt.Sprintf("/settings#group-%v", id), http.StatusFound)
 }
 
@@ -48,11 +61,22 @@ func DeleteGroup(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	group, err := database.GetGroup(int(id))
+	if err != nil {
+		ShowError(w, http.StatusInternalServerError, err)
+		return
+	}
+
+	username := sessions.GetUsername(r.Context())
+	if group.Username == nil || *group.Username != username {
+		ShowError(w, http.StatusForbidden, errors.New("you are not the owner of the group"))
+		return
+	}
+
 	if err := database.DeleteGroup(int(id)); err != nil {
 		ShowError(w, http.StatusInternalServerError, err)
 		return
 	}
 
-	// Redirect to settings.
 	http.Redirect(w, r, "/settings", http.StatusFound)
 }
